refactor(kakao): page 서버 컴포넌트로 변경
