+++ src/app/(route)/auth/kakao/callback/_components/index.ts
@@ -1,0 +1,1 @@
+export { default as KakaoContainer } from "./KakaoContainer/KakaoContainer";
